fix: update signals and legacy test file

+++ users/signals.py
@@ -13,16 +13,18 @@ from .models import UserProfile
 @receiver(email_confirmed)
 def update_user_email(sender, request, email_address, **kwargs):
     """Update the user's email when confirmed."""
-    email_address.set_as_primary()
-    EmailAddress.objects.filter(
-        user=email_address.user
-    ).exclude(primary=True).delete()
-    if request:
-        messages.info(
-            request,
-            f"{email_address} has been verified and set as primary email "
-            "address. Old E-Mail deleted."
-        )
+    # Only update if the email is verified
+    if email_address.verified:
+        email_address.set_as_primary()
+        EmailAddress.objects.filter(
+            user=email_address.user
+        ).exclude(primary=True).delete()
+        if request:
+            messages.info(
+                request,
+                f"{email_address} has been verified and set as primary email "
+                "address. Old E-Mail deleted."
+            )
 
 
 @receiver(email_added)
